More renames, fixing circular deps
